+++ CHANGELOG.md
@@ -10,6 +10,18 @@
 
 # Version History
 
+## v0.46.0
+
+- :tada: Add support for configurable web workers to allow custom levels of parallelism
+
+## v0.45.0
+
+- :rocket: Add EOT Character at end of streaming API calls
+
+## v0.44.1
+
+- :rocket: Add redirect from `/admin` => `/admin/index.html`
+
 ## v0.44.0
 
 - :tada: Allow creating and deleting bounds programatically